+++ source/src/SPCombatant.cs
@@ -11,6 +11,7 @@ namespace EnhancedBattleTest
     public class SPCombatant : EnhancedBattleTestCombatant
     {
         private readonly List<SPSpawnableCharacter> _characters = new List<SPSpawnableCharacter>();
+        private int _tacticLevel;
 
         public IEnumerable<SPSpawnableCharacter> SPCharacters => _characters.AsReadOnly();
 
@@ -19,20 +20,25 @@ namespace EnhancedBattleTest
         public override IEnumerable<BasicCharacterObject> Characters =>
             SPCharacters.Select(character => character.Character);
 
-        public SPCombatant(TextObject name, BattleSideEnum side, BasicCultureObject basicCulture, Tuple<uint, uint> primaryColorPair, Tuple<uint, uint> alternativeColorPair, Banner banner)
+        public SPCombatant(TextObject name, int tacticLevel, BattleSideEnum side, BasicCultureObject basicCulture,
+            Tuple<uint, uint> primaryColorPair, Tuple<uint, uint> alternativeColorPair, Banner banner)
             : base(name, side, basicCulture, primaryColorPair, alternativeColorPair, banner)
-        { }
+        {
+            _tacticLevel = tacticLevel;
+        }
 
-        public SPCombatant(BattleSideEnum side, BasicCultureObject culture, Tuple<uint, uint> primaryColorPair, Banner banner)
+        public SPCombatant(BattleSideEnum side, int tacticLevel, BasicCultureObject culture, Tuple<uint, uint> primaryColorPair, Banner banner)
             : base(GameTexts.FindText("str_ebt_side", side == BattleSideEnum.Attacker ? "Attacker" : "Defender"),
                 side, culture, primaryColorPair, new Tuple<uint, uint>(primaryColorPair.Item2, primaryColorPair.Item1), banner)
-        { }
+        {
+            _tacticLevel = tacticLevel;
+        }
 
         public static SPCombatant CreateParty(BattleSideEnum side, BasicCultureObject culture,
             TeamConfig teamConfig, bool isPlayerTeam)
         {
             bool isAttacker = side == BattleSideEnum.Attacker;
-            var combatant = new SPCombatant(side, culture,
+            var combatant = new SPCombatant(side, teamConfig.TacticLevel, culture,
                 new Tuple<uint, uint>(teamConfig.Color1, teamConfig.Color2),
                 teamConfig.Banner);
             if (teamConfig.HasGeneral)
@@ -60,7 +66,7 @@ namespace EnhancedBattleTest
 
         public override int GetTacticsSkillAmount()
         {
-            return 0;
+            return _tacticLevel;
         }
 
         public void AddCharacter(SPSpawnableCharacter character, int number)