+++ source/src/TroopConfig.cs
@@ -12,5 +12,11 @@
         {
             Character = CharacterConfig.Create(isMultiplayer);
         }
+
+        public TroopConfig(bool isMultiplayer, string id, int number, float femaleRatio = 0)
+        {
+            Character = CharacterConfig.Create(isMultiplayer, id, femaleRatio);
+            Number = number;
+        }
     }
 }